+++ ospray/include/ospray/OSPEnums.h
@@ -256,11 +256,13 @@ typedef enum
   OSP_FB_DEPTH = (1 << 1),
   OSP_FB_ACCUM = (1 << 2),
   OSP_FB_VARIANCE = (1 << 3),
-  OSP_FB_NORMAL = (1 << 4), // in world-space
-  OSP_FB_ALBEDO = (1 << 5),
+  OSP_FB_NORMAL = (1 << 4), // in world-space, for denoising
+  OSP_FB_ALBEDO = (1 << 5), // for denoising
   OSP_FB_ID_PRIMITIVE = (1 << 6),
   OSP_FB_ID_OBJECT = (1 << 7),
-  OSP_FB_ID_INSTANCE = (1 << 8)
+  OSP_FB_ID_INSTANCE = (1 << 8),
+  OSP_FB_FIRST_NORMAL = (1 << 9), // in world-space at 1st hit
+  OSP_FB_POSITION = (1 << 10) // in world-space
 } OSPFrameBufferChannel;
 
 // OSPRay events which can be waited on via ospWait()